+++ app/common/src/text.ts
@@ -112,7 +112,6 @@ interface PlaceholderOverrides {
   readonly logEventBackendError: [eventType: string]
 
   readonly subscribeSuccessSubtitle: [plan: string]
-  readonly assetsDropFilesDescription: [count: number]
 
   readonly paywallAvailabilityLevel: [plan: string]
   readonly paywallScreenDescription: [plan: string]
@@ -176,7 +175,8 @@ interface PlaceholderOverrides {
   readonly managingUserGroupX: [groupName: string]
   readonly planOverriddenToX: [planName: string]
   readonly 'manageLabelsModal.createLabelWithTitle': [labelName: string]
-  readonly assetsTableBackgroundRefreshIntervalOverridenToXMs: [ms: number]
+  readonly assetsTableBackgroundRefreshIntervalOverriddenToXMs: [ms: number]
+  readonly willUploadUpToXFileChunksAtOnce: [parallelism: number]
 }
 
 // This is intentionally unused. This line throws an error if `PlaceholderOverrides` ever becomes