+++ app/gui/src/providers/auth.ts
@@ -3,21 +3,17 @@ import * as backendModule from '#/services/Backend'
 import RemoteBackend from '#/services/RemoteBackend'
 import { BLACK_SQUARE_IMAGE_512PX } from '#/utilities/image'
 import type * as cognitoModule from '$/authentication/cognito'
-import {
-  featureFlagsForInternalTesting,
-  setFeatureFlags,
-  useFeatureFlag,
-} from '$/providers/featureFlags'
+import { useFeatureFlag } from '$/providers/featureFlags'
 import { useZustandStoreRef } from '$/utils/zustand'
 import { Opt } from '@/util/data/opt'
-import { ToValue } from '@/util/reactivity'
+import { proxyRefs, ToValue } from '@/util/reactivity'
 import { useToast } from '@/util/toast'
 import * as sentry from '@sentry/vue'
 import * as vueQuery from '@tanstack/vue-query'
 import { createGlobalState } from '@vueuse/core'
 import * as detect from 'enso-common/src/detect'
 import invariant from 'tiny-invariant'
-import { computed, inject, proxyRefs, toRef, toValue, watchEffect } from 'vue'
+import { computed, inject, toRef, toValue, watchEffect } from 'vue'
 import { createStore } from 'zustand'
 import { persist } from 'zustand/middleware'
 import { useBackends } from './backends'
@@ -251,12 +247,6 @@ function createAuthStore(
   gtagHooks.gtag('set', { platform: detect.platform(), architecture: detect.architecture() })
   gtagHooks.gtagOpenCloseCallback(gtagEvent, 'open_app', 'close_app')
 
-  watchEffect(() => {
-    if (userData.value?.type === UserSessionType.full && userData.value.user.isEnsoTeamMember) {
-      setFeatureFlags(featureFlagsForInternalTesting())
-    }
-  })
-
   const effectiveUserData = computed(() => {
     const intermediate =
       userData.value?.type === UserSessionType.full && planOverride.value != null ?
@@ -273,8 +263,7 @@ function createAuthStore(
   return proxyRefs({
     refetchSession,
     session: effectiveUserData,
-    waitForSession: () =>
-      sessionData.waitForSession().then(() => queryClient.ensureQueryData(usersMeQueryOptions)),
+    waitForSession: () => sessionData.waitForSession().then(() => usersMeQuery.suspense()),
     setUsername,
     isUserMarkedForDeletion,
     isUserDeleted,