+++ lib/scala/logging-config/src/main/java/module-info.java
@@ -1,8 +1,14 @@
+import org.enso.logging.config.systemlogger.ContextLoggingViaSlf4j;
 import org.enso.logging.config.systemlogger.SystemLoggerViaSlf4j;
 
 module org.enso.logging.config {
   requires org.slf4j;
+  requires java.logging;
   requires typesafe.config;
+  requires org.enso.logging.utils;
+  requires org.enso.engine.common;
+  requires static org.graalvm.polyglot;
+  requires static org.openide.util.lookup.RELEASE180;
 
   exports org.enso.logging.config;
 
@@ -10,4 +16,6 @@ module org.enso.logging.config {
 
   provides java.lang.System.LoggerFinder with
       SystemLoggerViaSlf4j;
+  provides org.enso.common.ContextLoggingConfigurator with
+      ContextLoggingViaSlf4j;
 }